Version 1.3.2 - major bug fix
> Strings ending in a back-slash ( \ ) are serialized the old way.

I knew it was a good idea to keep it, hah.
This is actually the best way, because I don't need to escape
backslashes. If there is only one backslash (which happens to be at the
end), then my method is equivalent to escaping the backslash, in size,
but actually faster to read.
I got the best in both worlds going here.